backend/requirements.txt
fastapi
uvicorn
pymupdf
faiss-cpu
sentence-transformers
openai
python-dotenv
python-multipart
